requirements.txt
flask>=2.3.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
celery>=5.3.0
redis>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
loguru>=0.7.0
pyyaml>=6.0
psycopg2-binary>=2.9.0
requests>=2.31.0
tushare>=1.2.0
akshare>=1.10.0
python-multipart>=0.0.6
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9
cryptography>=42.0.0
prometheus-client>=0.19.0
structlog>=24.1.0
